requirements.txt
# Web framework
flask>=3.0.0
werkzeug>=3.0.0
gunicorn>=21.2.0

# AI / ML
openai-whisper>=20231117
transformers>=4.40.0
torch>=2.2.0
torchaudio>=2.2.0

# Utilities
python-dotenv>=1.0.0
